* Updated firmwares to the latest
* Refactoring
* More than 255 tokens is possible to add into app now

+++ totp/scenes/token_menu/totp_scene_token_menu.c
@@ -18,7 +18,7 @@ typedef enum { AddNewToken, DeleteToken, AppSettings } Control;
 
 typedef struct {
     Control selected_control;
-    int16_t current_token_index;
+    int32_t current_token_index;
 } SceneState;
 
 void totp_scene_token_menu_init(const PluginState* plugin_state) {
